Move toaster to compose, make settingsview work as screen
